common, login, game: remove ctype.int

+++ game/game/packets/char_create_ok.py
@@ -8,4 +8,4 @@ from .base import GameServerPacket
 @dataclass(kw_only=True)
 class CharCreateOk(GameServerPacket):
     type: ctype.int8 = field(default=25, init=False, repr=False)
-    result_ok: ctype.int = field(default=1, init=False, repr=False)
+    result_ok: ctype.int8 = field(default=1, init=False, repr=False)
